# Conflicts:
#	game-admin/game-admin-controller/src/main/java/com/fulu/game/admin/controller/ImController.java
#	game-app/game-app-controller/src/main/java/com/fulu/game/app/controller/ImController.java
#	game-app/game-app-controller/src/main/java/com/fulu/game/app/controller/UserController.java
#	game-core/game-core-service-impl/src/main/java/com/fulu/game/core/service/impl/ImServiceImpl.java
#	game-core/game-core-service/src/main/java/com/fulu/game/core/service/ImService.java
#	game-play/game-play-controller/src/main/java/com/fulu/game/play/controller/ImLogController.java

+++ game-core/game-core-dao/src/main/java/com/fulu/game/core/dao/DynamicCommentDao.java
@@ -14,7 +14,7 @@ import org.apache.ibatis.annotations.Mapper;
  * @date 2018-08-30 11:21:42
  */
 @Mapper
-public interface DynamicCommentDao extends ICommonDao<DynamicComment,Long>{
+public interface DynamicCommentDao extends ICommonDao<DynamicComment,Integer>{
 
     List<DynamicComment> findByParameter(DynamicCommentVO dynamicCommentVO);
